fix(keybindings): Fix 'command -tabout not found' msg
Previous pull request had a keybinding command -tabout which worked
when it was in vscode 'keybindings.json' file but caused the TabOut
extension to error out because it couldn't find '-tabout' command.

This commit removes that portion of the keybinding. It has been tested
in debug mode in VScode and appears to function as expected.

Resolves: #14